+++ tauri-wallet/src-tauri/src/operations/mod.rs
@@ -1,0 +1,5 @@
+pub mod account;
+pub mod admin;
+pub mod bond;
+pub mod delegate;
+pub mod send;
