+++ tauri-wallet/src/types/rust/account.ts
@@ -1,0 +1,8 @@
+import { Denom } from "./denom";
+
+export interface Account {
+  contract_address: string;
+  client_address: string;
+  denom: Denom;
+  mnemonmic: string | null;
+}
